renamed api-docs.js to api_docs.js
